basic04: assignment 2 solution that _doesn't_ work
The "solution" for this requires either a complete rewrite of
`load_bpf_and_xdp_attach` from the library, or to just use
`cfg.reuse_maps`. I tested with the latter, and it works, but that's not
really "my" work. I opted to not do the former, because it seems like a
waste of time.

This solution doesn't work. I believe this is because we load the bpf
program _before_ we tell the object to reuse the existing maps, so the
actual loaded program uses its own maps and the reuse call doesn't
affect it.

Signed-off-by: Tal Shorer <shorer@amazon.com>